semana04/dia3/requirements.txt
cloudinary==1.26.0
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
whitenoise==5.3.0

